pyproject.toml
[project]
name = "django-starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ratelimit>=4.1.0",
    "django>=5.2.1",
    "djangorestframework>=3.16.0",
    "redis[hiredis]>=6.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "python-decouple>=3.8",
    "pydantic>=2.11.4",
    "drf-spectacular>=0.28.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.0",
    "django-redis>=5.4.0",
    "pillow>=11.2.1",
    "django-filter>=25.1",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "daphne>=4.1.2",
    "django-guardian>=3.0.0",
    "twisted[http2,tls]>=24.11.0",
    "django-mptt>=0.17.0",
    "django-countries>=7.6.1",
    "django-rest-framework>=0.1.0",
    "az-iranian-bank-gateways>=2.0.12",
    "qrcode>=8.2",
    "django-cors-headers>=4.7.0",
    "psycopg2-binary>=2.9.10",
    "django-encrypted-model-fields>=0.6.5",
    "pycrypto>=2.6.1",
    "sentry-sdk>=2.28.0",
    "django-celery-beat>=2.8.1",
    "celery>=5.5.2",
    "django-celery-results>=2.6.0",
    "django-debug-toolbar>=5.2.0",
    "django-silk>=5.3.2",
    "drf-nested-routers>=0.94.2",
    "tblib>=3.1.0",
    "django-extensions>=4.1",
    "psutil>=7.0.0",
]
